Changed ScanLine Intersection algorithm, now the points near edge can be filled.
1. In previous implementations, I use equations to calculate the X value
of intersection. But we draw polygon edge with Bresenham, which means
that the X value we calculated is correct in math, but not on the
screen. Points near the edge are not filled correctly. Now, I directly
traverse the points on the line and find the first point which Y value
equals scanline y value. Now the points near edge can be correctly
filled.